iotivity: fix memory leakage in HandlePostRequest()
This patch fixes the memory leakage in HandlePostRequest() function.
HandlePostRequest() intends to free the memory of 'cred' list only
when the member of 'cred' list is attached to gCred list successfully.
But, the member of 'cred' list should be freed even if it is not
attached to gCred when it is the empty member. This patch adds the
codes which frees the member of 'cred' list when it is empty.

This patch assumes that 'cred' list created by CBORPayloadToCred()
is maintained even after HandleNewCredential() is executed. But,
'cred' list can be damaged because the pointer of the 'cred' list's
member is passed to HandleNewCredential() and it is modified while
it is attached to gCred. It is iotivity bug and it should be fixed.

This patch works correctly if the number of 'cred' list members is
one. But, the memory leakage can be left in case of multiple members
of 'cred' list because of above bug.

Change-Id: I80c316985bbf7077f2d542090feed27f9cf669a4
Signed-off-by: Wonsang Ryou <wonsang.yoo@samsung.com>