external/wakaama-client: fix release of resolved addresses pointer
The address resolving information returned by 'getaddrinfo' was
release using 'free' instead of dedicated 'freeaddrinfo'. This
seemed to work with the previous version of the LWIP stack but
it causes a crash since we migrated to latest LWIP.

This patch fixes this crash by using dedicated API for releasing
the resolution info.

Change-Id: If9a681f90e1c2fd7fd862fb2a5f1a271dcdcc622
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>