Add a sample app of RT-OCF framework
RT-OCF is a lightweight D2D framework based on OCF specification.
The added sample app is a server example.
To run this sample app, RT-OCF source code needs to be downloaded from public RT-OCF github and be placed at 'TizenRT/external'.
To see 'How to start' in detail, please refer "README.md" or the main page of the RT-OCF github.