framework/iotbus: make wrappers for each component, i2c, gpio, pwm, spi, and uart
when closing context, context should be set as NULL with cleaning the all resources.
but now context is NULL in function only because user passes the context as parameter to close function.
It means passed value is value, not address.
So we make the wrapper of iotbus structure to get the address of it.