examples/artik_examples: fix certificate verification in cloud example
The cloud example would configure TLS to always verify the root CA
during handshake. However when the Secure Element is used to connect to
ARTIK Cloud, the HTPPs endpoint presents a self-signed certificate that
cannot be verified against ARTIK Cloud's usual root CA.

This patch disables server certificate verification when the "use_se"
parameter is passed to the "connect" and "dm" subcommands.

Change-Id: Id77eedefb0875e5ad4394f7fa8ce73dd8a161c56
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>