iotjs: Relocate TizenRT's kconfig to align upstream
This file have been upstreamed in config's subdir so we align to it.

Relate-to: https://github.com/Samsung/iotjs/pull/1529

Now, The module is easier to update,
by just replacing iotjs subtree with upstream tree.

Note for later, for developer convenience and to ease upstreaming works
(tracking changes, logs etc) the folder can be replaced by submodule,
like other deps in Tizen:RT,
but at the cost of losing all changes occurred in after version import.

Change-Id: If4a7d71d862745173f38d009924f3e4ce98cb6d0
Bug: https://github.com/Samsung/TizenRT/pull/1277
Signed-off-by: Philippe Coval <philippe.coval@osg.samsung.com>