artik05x: enable thumb-mode
If the user enables Thumb mode, the user can see the effect of reducing
the size of the binary. However, because we can not guarantee that
performance will improve, we have left it as an option. Users who want
Thumb mode can change the following config in menuconfig.

For the verification of thumb mode, turn on CONFIG_ARCH_THUMB.

CONFIG_ARCH_THUMB=y

Change-Id: I0da6a6fa6e463588c94c8c75a0c5cc7285f86fb7
Signed-off-by: EunBong Song <eunb.song@samsung.com>
Signed-off-by: Junhwan Park <junhwan.park@samsung.com>