dhcpc: optimize recv timeout for socket
This commit is to optimize the socket receive timeout for dhcp client.
Currently it has 10 seconds for recv timeout. As DHCP DISCOVER is sent
as broadcast packet, it might be lost in the air with high possibility.
It's relatively big value because it can resend the DISCOVER message
again after recv timeout.

Change-Id: I320bd4d8405978e96a880066c8f826238e4cd39b
Signed-off-by: Changsoon Kim <cs0424.kim@samsung.com>