external/iotjs: get build options, target-arch, board from Kconfig
When we try to build IoTjs, we should set --target-arch and --target-board
options on Makefile. They are fixed to arm and artik05x but they cause
difficult to find and wrong operation on IoTjs.
When TizenRT is built, CONFIG_ARCH and CONFIG_ARCH_BOARD are set and
they shows the information of selected board. User doesn't need to
set them in iotjs again.
This commit makes iotjs using those configurations.