netutils/webserver: fix double close socket in webserver
The https webserver closes the socket after client request processing
and closes the socket again when tls release. This is not a problem with
a single client. However, if the socket number which closed by
termination of client connection and the socket number of the newly
connected client is same in multi-client environment, the new socket is
closed.

Change-Id: I5ca3929196e92fa6b9c9d5110a9a0ccc66666252
Signed-off-by: Heuisam Kwag <heuisam@samsung.com>