sem: add a flag to show initialization and check it on sem APIs
1. POSIX standard describes return value as shown below:
Upon successful completion, a value of zero shall be returned.
Otherwise, a value of -1 shall be returned and errno set to indicate the error.
  But our sem APIs do not check sem initialization and do not return ERROR.
2. Without sem_init, almost sem APIs work. But it can cause abnormal
  operation.

The sem_s structure already has the flag element inside
and first bit is an inheritance flag. Let's use second bit as
initialization flag and check it at first operation of sem APIs.