mm_heap: allocator instrumentation
Allocator instrumentation needed for creating redzones around objects
located on heap. In case of mm_heap allocator, we poison (mark as a
redzone) mm_allocnode_s structure for each allocation for detecting
overflow and underflow of memory chunk. And we poison whole memory
chunk with mm_freenode_s structure if it is free and put it in
quarantine for detecting use-after-free bugs.

Signed-off-by: Roman Lakeev <r.lakeev@samsung.com>