webclient: fix authentication mode override
When the caller application provides a specific SSL authentication
mode, this latter is overridden by the webclient library setting it
to a hardcoded default value (disabled by default). This patch
moves the call to the configuration function to only happen when
the webclient library does the mbedtls configuration initialization,
and not when the initialization is done by the calling application.

This patch also removes the "auth_mode' of the SSL configuration
structure, this one is no longer used since it has been replaced
by the hardcoded value.

Change-Id: Iacbe30bde18fee168c5c76109e64856a14b6f357
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>