framework/st_things : Changes OCRepSet/Add*AsOwner to OCRep*.
OCRepSet/Add*AsOwner function doesn't allocated memory
when build the response payload.
So, user should allocate the memory for payload but user can't free.

In order to ignore this issue , We can usie OCRep* instead of OCRepSet/Add*AsOwner.
And, we should care of the pointers to avoid memory leak.