net/lwip: take account any address for certain operations
The patch consists to resolve problem we can meet when we want
to bind a socket after retrieving IP address and port with
getsockname for example. Indeed, if we have bound a socket with
AF_UNSPEC, IPv6 frames will be handled before to be mapped into
IPv4 frames if a client has sent a IPv4 request. We noticed that
getsockname doesn't verify if IP address is any and can make the
future socket operations not working.

Change-Id: I082dddf4675a1bab23290ad8f14b71a713b1adee
Signed-off-by: Romain Cancilliere <romain.c@samsung.com>