cloud: Make cloud command synchronous
The cloud dm connect and cloud dm disconnect commands need to be executed
in the same task. Because the cloud dm connect creates a new thread and
the cloud dm disconnect performs a pthread_join on this thread before
destroying a mutex. In asynchronous mode, pthread_join failed and
the shared mutex is destroyed before the lwm2m thread unlocks the mutex.

Change-Id: Id0cc61989448dfb0aa34cd87269bb8024e90264c
Signed-off-by: Camille Bégué <c.begue@samsung.com>