build: add new flag to disable git submodules update
When building an image with the ARTIK SDK enabled, an update
of the git submodules are triggered automatically upon calling
'make'. This patch allows the user to disable this behavior by
setting the NO_SUBMODULES_UPDATE environment variable. This is
useful when developers make changes in the SDK modules and do not
want their local changes to be overridden by the submodule update.

Change-Id: I6d01866d242a1c6dd79019ee60e97987f6308587
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>