apps/testcase/systemio: fix compilation errors on UTC
When CONFIG_SERIAL_TERMIOS is disabled, below errors are shown:

/os/../build/output/libraries/libapps.a(utc_uart.o): In function `utc_systemio_uart_set_baudrate_p':
/apps/examples/testcase/ta_tc/systemio/utc/utc_uart.c:59: undefined reference to `iotbus_uart_set_baudrate'
/apps/examples/testcase/ta_tc/systemio/utc/utc_uart.c:59: undefined reference to `iotbus_uart_set_baudrate'
/os/../build/output/libraries/libapps.a(utc_uart.o): In function `utc_systemio_uart_set_baudrate_n':
/apps/examples/testcase/ta_tc/systemio/utc/utc_uart.c:65: undefined reference to `iotbus_uart_set_baudrate'
/apps/examples/testcase/ta_tc/systemio/utc/utc_uart.c:65: undefined reference to `iotbus_uart_set_baudrate'
...

Above functions of IoTbus UART framework are built when CONFIG_SERIAL_TERMIOS is enabled.
This commit make a conditional to that config.

When -Wno-implicit-function-declaration is not set in Make.defs, system I/O UTC will
have compilation errors because of missing prototype of each system I/O UTC function.

This commit adds prototype of all of functions which calls from other file.