artik_onboarding: add support for custom SDR DTID
Currently the onboarding service would use a hardcoded
value to know if the currently used device type ID was
SDR-enabled and needed proper secure path to connect to
ARTIK Cloud. This means that developers had to modify the
code and rebuild the firmware in case they changed the
default DTID for a SDR-enabled one, which can be done
dynamically.

This patches makes this process dynamic instead, by detecting
at onboarding time if the DTID is secure or not, based on
the APIs called by the mobile apps which are different
between the SDR and non-SDR onboarding paths.

Change-Id: I9d0d21e63b2d314322b697aff725d359a240f995
Signed-off-by: Camille Bégué <c.begue@samsung.com>