libc/math: Fix wrong implementation of err*() functions
Implementation of err*() functions were wrong as sign and
other logical cases were not handled correctly.

The implementation is changed and all cases are taken care of,
so that output of erf*() functions is similar to output of
erf*() functions in other math libraries.

Note: The output is verified with glibc math library of linux.

Signed-off-by: Lokesh B V <lokesh.bv@partner.samsung.com>