external/wakaama-client: fix memory issues
This patch is the result of using static and dynamic analysis tools
on wakaama-client to detect and fix memory leaks, unterminated strings,
null pointer dereferencing, unreachable code, and value range issues.

This is a first iteration that shows significant memory usage
improvement under the scope of usage of the ARTIK SDK cloud DM
example and ARTIK onboarding service. It is not an exhaustive cleanup
of all the potential issues detected by the analysis tools, for which
several more iterations will be required.

Change-Id: I3cfca795c169d0e2a67f3ab59986556db9b02cf5
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>