examples/artik_onboarding: add retry mechanism for the NTP client
Under flaky connectivity conditions, the NTP client sometimes fails
to reach the server an exits on error without setting the date, which
may cause other services to fail if the system date is invalid.

This patch adds a retry mechanism to work around this issue by
restarting the NTP client until it succeeds to set the date. It also
increases the NTP response timeout to 10 seconds to minimize the
appearance of the issue on flaky connections.

Change-Id: Ic455b7fafb82922255b44cacd676f4091a6b5865
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>