netutils/websocket: fix infinite loop
When disconnecting the websocket, the close function tries to
send a close packet and waits forever for the server to close the
connection and for the websocket to reach the STOP state.

Under certain circumstances - e.g. the Wi-Fi association is still on
but the access point does not relay IP packets to the internet - the
websocket never reaches the STOP state, the loop iterates forever and
blocks the calling thread.

This patch fixes this issue by specifying a maximum number of iterations
of the loop until which the loop exits reporting an error. This allows
calling application to avoid remaining blocked while being able to take
appropriate actions in case such situations happen.

Change-Id: Icd4f7f9d4ea730aef8b5be088469bc2d7c15d770
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>