apps/sysdbg: Fix the logic of handling the args lg
Sometimes, when args[2] is not passed in TASH
ex:  TASH>>sysdbg help
length of args[0] is 6
length of args[1] is 4 and
length of args[2] is 32 , even when args[2] is not passed at all
This causes program to enter wrong logic and always returns error.

This patch address this issue by checking the args length based on
number of arguments passed.

Signed-off-by: pradeep.ns <pradeep.ns@samsung.com>