Add new defconfigs artik053/hello_mem_opt artik053/tc_mem_opt
This patch contains two defconfigs to support little profile in
TizenRT. In this profile, some of the kernel features are disabled
such as TASH, HPWORK etc. and some of them are optimized.
artik053/tc_mem_opt can be used to run the test cases which can be
run with this optimization.