kernel/environ: Fix an error in the duplication of the child tasks environment
Fix an error in the duplication of the child tasks environment in the special case
where the parent's environment was created, but then all of the variables were unset.
In that case, there is still an allocation in place but the size of the allocation is zero.
This case was not being handled correctly when a child task attempts to create its environment
and inherit the zero-size parent environment. Noted by Anthony Merlino.

This patch is back-ported from Nuttx.

Signed-off-by: Lokesh B V <lokesh.bv@partner.samsung.com>