netutils/webserver: fix buffer overflow
When handling HTTP headers, the webclient library does not
check their length before copying the value sring into a
statically allocated string. In cases where the length of the
header's value would be longer than the predefined destination
string length, this would result in a buffer overflow. This
kind of situations can commonly happen when targeting websites
exposing the "Content Security Policy" header, which usually contain
a long list of approved URIs.

This patch fixes the issue by trimming the header string to only copy
the allowed length. In such cases, the calling application should
be careful and handle headers knowing that they could be incomplete.

Change-Id: I153e1012f20cedbf7cc0c2a6abbc4a209991ecac
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>