examples/smart_test: modify to add options for verbose
Currently, running the this test will be print many logs. This means
that the processing time of this example is delayed because of the time
it takes to print logs. This commit modifies to add option for verbose.
The default value for this option is '1'. no logs will be print if you
enter the value as '0'.

-----------------------------
TASH>> smart_test -c 0 -s 0 -w 1 -e 32 -r 64 -t 400 -l 2000 -v 0
/mnt/log

Creating file /mnt/log for write mode
Writing test data.  2000 lines to write

Done.
Performing 1 random seek with write tests

Pass
TASH>>
-----------------------------

Change-Id: I642be3ebc0db8bd80bc97139f0ce845c166f9ad6
Signed-off-by: Jihun Ahn <jhun.ahn@samsung.com>