network/tc: Add set socket option
add logic to reuse sockets of the same address.

If you do not use this option, the socket will not open when reusing the socket. The reason is that when you close the socket you used before, it will not close soon with the reuse in mind and it will be delayed closing.

I deleted the network_tc that was not needed in the past commit.If there was a large number of network_tc, I had enough time to close the previously used socket, but now I run the following tc The socket can not be created on the same address because the socket is not closed before the socket is opened.