configs/artik05x: change script to python executable
Currently, this script file that is download binary to flash memory was
created by bash shell format. In additional, This file is made to be
depends on GNU/linux syntax, but macos does not fully support GNU
syntax(it uses BSD syntax). Therefore, it might not run on Windows, or
macos.

This commit changes the download script format to python.

Change-Id: I945386d209a51e971ddecf92288f45b606639843
Signed-off-by: Jihun Ahn <jhun.ahnn@gmail.com>