external/libopus: Add new stack allocation mode
Opus requires one of VAR_ARRAYS, USE_ALLOCA or  NONTHREADSAFE_PSEUDOSTACK
be defined to select the temporary allocation mode.
But it takes large size stack or fixed size heap (non-threadsafe),
now we add a new mode: USE_SHAREDPTR, it allocates adaptive heap size
for VLAs and automatically release the memory when it's not used.
Actually, we use the feature of std::shared_ptr to allcate VLAs from
heap instead of stack with auto release.