configs/artik05x: mv ld script to common directory
The each artik05x board(artik053, artik053s and artik055s) use the same
LDSCRIPT file(i.e. flash.ld). So, this file doesn't have to be in each
board config directory. This commit modifies to moves each LDSCRIPT to a
common path(configs/artik05x) and to use it.

Change-Id: I0852a759453dacb08711ee7b41d88350b5907b89
Signed-off-by: Jihun Ahn <jhun.ahn@samsung.com>