kernel/task : add dynamic support for atexit & onexit
There is a limitations in the current implementation that
only a single atexit/onexit function can be registered unless
CONFIG_SCHED_ATEXIT_MAX/CONFIG_SCHED_ONEXIT_MAX/ defines a
larger number.

If the MAX number is not sufficient, then the subsequent
callback registration would fail. It is also hard to
estimate the MAX number of callback functions.

To overcome this limitation this patch removes the support of static
array of callbacks and keeps the callbacks in the dynamically allocated
space at runtime using dqueue.

Signed-off-by: Manohara HK <manohara.hk@samsung.com>
Signed-off-by: Yashwanth <v.yashwanth@samsung.com>