netutils/dhcpd: introduce a way to show lease pool info
This commit is to implement a way to show DHCPD's lease pool
information.
We need to check current lease pool information off and on for debugging
purpose.
The following is how to use and example of the result.

* How to use and example
TASH>>dhcpd status

dhcpd status : running
-----------------------------------
Lease Pool Information
-----------------------------------
This is for only debugging purpose.
Please note that it doesn't guarantee thread-safe with dhcpd task.
Current Time=1262304112
0 | Alloc: 1 | IP=0xc0a80a02 | MAC=28:6d:97:40:94:7c | Expired
at=1262304109 (expired)
1 | Alloc: 1 | IP=0xc0a80a03 | MAC=f8:e6:1a:31:0d:86 | Expired
at=1262304154 (42 sec left)
2 | Alloc: 0
3 | Alloc: 0
4 | Alloc: 0
5 | Alloc: 0

Change-Id: I0fe09527b1f5add9f42a0824778607d303529e2c
Signed-off-by: Changsoon Kim <cs0424.kim@samsung.com>