[DRIVER] Update scan error handling
If the scan req is not sent successfully, then the scan fails.
In such a case a scan completion event is not notified to upper layer
where error code < 0.

This could cause the upper layer to wait for SCAN event which
is not sent and eventually might cause the API to get stuck.

Send a scan completion event in all cases of failure.

SCSC-Bug-Id: S17SWF-819

Change-Id: I0617784677ef49324d313c37297fe33257c4d437
Signed-off-by: Anoop Nair <anoop.nair@samsung.com>