os/tools: add download scripts for Linux and Windows
"make download" automatically flashes TizenRT binary into a board.
But we do not public support Windows so that "make download" is
not working and it is difficult to execute "make" on it.
To give Windows user convenient, download.bat is added.
But it deos not perfectly provide same usages with "make download".

To keep same usages with download.bat, download.sh is added as well.
But it just executes "make download".

Signed-off-by: sunghan-chang <sh924.chang@samsung.com>