configs/artik05x: fix eeprom configuration
Currently, After writing FTDI eeprom, JTAG doesn't connect except
openocd. Because the driver of ft2232H's channel A is overwritten to
UART. In this case, ADBUS2(set to TDO) pin is fixed to logic-1. To use
JTAG device, you have to specify the driver in eeprom configuration
(in this configuration suggested 'cha_type'). So, This commit fixes
eeprom configuration to support using JTAG device.

Change-Id: I1984d85e78036c6f846f52a06fa98e75768f59bc
Signed-off-by: Jihun Ahn <jhun.ahn@samsung.com>