iotjs: copy the IoT.js static libraries to the external iotjs folder
Currently, all the iotjs static libraries are always searched in the embedded
iotjs folder (external/iotjs). If the IOTJS_ROOT_DIR environment variable is
set (to use a custom iotjs), the created libraries are on different path, that
makes an error during the build.
This patch modifies the iotjs makefile to copy the created iotjs static libraries
to the embedded iotjs folder.