testcase/task_manager: Modify timeout param for task_manager_start to get return
If timeout param sets to value, it means to wait timeout ms. But in some cases, it cannot receive the return properly.
In our TC, after unregistration, try to start the unregistered app and try to check the return is TM_UNREGISTERED_APP or not.
For checking the return properly, timeout should be set to TM_RESPONSE_WAIT_INF.