build: add board specifc operation to make a board-specific binary on build time
The board-specific binary was made at download step.
But it is not a download step and some board uses toolchain to do it
like strip command in Cypress.
It is not working in Docker environment.
Let's call board-specific operation after making TizenRT binary
and define it at each board config file name Make.defs.

Signed-off-by: sunghan-chang <sh924.chang@samsung.com>