wpa_supplicant: do not add or update an hidden SSID
This commit is not to add or update an hidden SSID during scan.
T200 Wi-Fi fimrware is getting the scan results from Beacon frames as
well as Probe Response. Hidden SSID never used by an application since
it doesn't have identity to deal with. So, it doesn't need to be added
at all.

Change-Id: I332224b4f271780b9e248a5044241d68d5a83112
Signed-off-by: Changsoon Kim <cs0424.kim@samsung.com>