external/wakaama-client: add activation of Raw Public Key extension in SSL
When the client want to connect to an endpoint and the server
sends a Server Hello with client_certificate_type and
server_certificate_type extensions, the client will want to parse
these extensions. The problem is the extensions are disabled
in client-side and will send an alert message. So we active these
parameters in client-side to resolve this.

Change-Id: I01c65a3fbcc1e727ca059801abf52553b215e5db
Signed-off-by: Romain Cancilliere <romain.c@samsung.com>