external/wakaama-client: verify if status of registration is "pending"
When we wanted to connect with "cloud dm", we received a disconnection
event. It's because we don't verify if the status of registration is set
to STAGE_REG_PENDING and also, the error was set by default to
LWM2M_CLIENT_DISCONNECTED and was returned by the LWM2M client service.

Change-Id: I74f1ad715c3f19dd41cc81389c5eee21d19dc705
Signed-off-by: Romain Cancilliere <romain.c@samsung.com>