framework/task_manager: Modify internal function (taskmgr_getinfo_with_name).
According to the type of task/thread (built_in task, task, pthread), the lists that stores name of task/thread are different.
Therefore, an invalid referencing can occur which can result in an assert problem.
The procedure of selecting the lists to find the name of task/thread was added according to the type.