framework/task_manager: Add procedure to change the group of tcb.
Previously, when pthread is created by task manager (when task_manager_start() is called), the group of pthread's tcb is set as the group of task manager.
The group of pthreads' tcb needs to be set as the group of the caller.
Otherwise, this can leed to a inappropriate result.

Therefore, additional procedures were added to prevent this problem.
After a pthread is created by task manager, the group of pthread's tcb is detached from the task manager.
Then, the group of pthread's tcb is joined to the group of caller.
Furthermore, the ppid of pthread's tcb is changed into the pid of caller