Get the current location from script file path on shell scripts
Many sh files get a current path using PWD command. But it
is changed where you run it.
For example, if you run a shell script on aa/bb folder, PWD is aa/bb.
But if you run it like bb/xx.sh on aa folder, PWD is aa.
The location of executing shell script affects working of shell script.

This commit gets the current path from location of shell script which
is fixed. Wherever you run it, it works.

Signed-off-by: sunghan-chang <sh924.chang@samsung.com>