external/artik-sdk: fix deadlock in LWM2M connection
There was an unreleased mutex situation that would happen
when reading a LWM2M resource after being disconnected by
an external factor. This would lead to a deadlock when trying
to read a resource twice in a row. This patch properly
releases the mutex when needed to avoid this situation.

Change-Id: Ie4d341d972549ef524deece87d05b3478175f95f
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>