Resize number_buffer array in cJSON.c to solve the buffer overrun issue.
iperf app uses cJSON library sending and receiving app results including the bytes sent.
When you send the large bytes to a server, especially using 5GHz WiFi,
the 'number_buffer' array, which is temporary used to print the number in cJSON, can be overrun.
That will cause sending results error, and you will see the log : "error - unable to send results".
The size of 'number_buffer' array was set to 27 to solve the error, that is enough to get the large bytes.
If you encounter the same problem, however, check the 'number_buffer' array size again in cJSON.