kernel/init: allow CONFIG_USER_ENTRYPOINT to be NULL
Originally CONFIG_USER_ENTRYPOINT can not be undefined with
normal usages. But someone modifies defconfig directly (actually
it is wrong usages). For that case, mkconfig.c has a conditional.
If it is undefined, mkconfig.c defines it as "main".
But we can't guarantee it is right so that this commit removes it
to be a compilation error.

Additionally TizenRT allows to execute an application from TASH.
That means someone can use TizenRT without user entry point
which is started automatically at bootup.
When CONFIG_TASH is enabled, this commit allows that
CONFIG_USER_ENTRYPOINT is undefined.
At that case, TizenRT just starts TizenRT's services only.
User application should be launched from TASH.

Signed-off-by: sunghan-chang <sh924.chang@samsung.com>