utils/kdbg: fix bug and change strtok to strtok_r for tokenizing stat info of proc
1. Fix bug about loop iteration
The number of stat info is PROC_STAT_MAX.
A function for tokenizing should be called (PROC_STAT_MAX-1) times.

2. Change strtok to strtok_r
Kdbg commands use strtok for tokenizing stat info of proc.
But strtok is not thread-safe because it use a global variable, g_saveptr.
So it should be changed to strtok_r.