Fix the network testcase 'tc_net_ioctl_fionread_n'
The ioctl command FIONREAD is defined twice both in tinyara/fs/ioctl.h and net/lwip/sockets.h,
and lwip_ioctl fuction uses the command defined in tinyara/fs/ioctl.h.
We disable FIONREAD in net/lwip/sockets.h to avoid conflict.

The testcase 'tc_net_ioctl_fionread_n' was passed because the parameter FIONREAD which is defined in net/lwip/sockets.h didn't match that used in lwip_ioctl function.
So we change this testcase to a positive case because it works well with FIONREAD command.