arch/arm: fix invalidate branch predictor
The "mcr CP15_BPIALLIS(r0)" instruction is used for invalidating entire branch
predictor. This is the generic code and can be called on any armv7-r
architecture based CPU's.

It is a problem, if this instruction is called in uni processor configuration.
Because, BPIALLIS (c7, 0, c1, 6) instruction is only added as part of the
"Multiprocessing Extensions"(As per ARM® Architecture Reference Manual
/ARMv7-A and ARMv7-R edition)

Hence this instruction should be under SMP configuration.
In non-SMP configuration this instruction could become undefined.

Signed-off-by: Manohara HK <manohara.hk@samsung.com>