arch/armv7-r: Modify MPU configuration to accept region number
Modify MPU configuration to accept region number.
When app binary separation is enabled, the mpu configuration api's are called multiple
times during load or restart of an application. The current implementation would assign
new mpu region each time. But we want to change an existing region each time the api is
called. Hence, we modify the api to take the region number and make changes to the region.
Also, modify previous calls to this api to take the region number as arguement.

Signed-off-by: Kishore SN <kishore.sn@samsung.com>