apps/examples/testcase: Add dependency configuration filesystem tc
I made filesystem TC dependent on the !BUILD_PROTECTED setting.

reason1 : The file system TC has an API that calls the kernel context. Therefore, it is disabled in the BUILD_PROTECTED state.

reason2 : Because the file system TC is created for testing purposes, not for actual code, disabling BUILD_PROTECTED does not cause any problems