board/common: change a location of sub-title in Kconfig
A sub-title of flash partition was in board/Kconfig and it
prints sub-title before sourcing board/common/Kconfig.
Because of this, when it is not available, only sub-title
is come up without any config as shown below:

(4096) Page size of external flash memory (in bytes)
    *** Board-Partition Options ***

To hidden it when it is not available, move it inside of
board/common/Kconfig and cover it using conditionals.

Signed-off-by: sunghan-chang <sh924.chang@samsung.com>