Fix implementation of User MM (kumm_xxx) API's
kumm set of MM API's are used by kernel to allocate memory from the
user heap. When APP_BINARY_SEPARATION is enabled, the heap object for
each app lies at the beginning of the app's RAM partition. When the app
is loaded, we will store a pointer to the app's heap object in its TCB's
and kumm API's will use this pointer for all memory management.

Signed-off-by: Kishore SN <kishore.sn@samsung.com>