apps/examples : add a memory fragmentation test
This app allocates and releases small and large memory segments in a mixed-up manner repeatedly
in order to intentionally fragment the heap memory.
'heapinfo' shows how the heap is fragmented in detail.
Note that fragmentation tends to become severe when the lifetime of small memory
is relatively longer than that of large memory.