Add support of abort cases in fault manager
In case of data/prefetch/undef aborts, in earlier implemenation
board used to hang.

In fault manager, we check whether the abort is in kernel or user
context. If it is in user context, application causing the abort
will be restarted and if it is in the kernel context, board will
be rebooted.

Signed-off-by: Drashti Parikh <p.drashti@partner.samsung.com>