Add MPU test for App separation scenario
During MPU test, the WIFI app and Micom app are running.
The Wifi app tries below tests:

	Read / Write from Micom app space
	Read / Write from kernel code space
	Read / Write from kernel data space

Each of these tests can be considered successful if it results in a MemManage fault.
If we dont want to execute the MPU test, we can just press enter key to continue normal execution of Wifi App.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>