os/arch/arm/src/imxrt: Configure FlexRAM banks for DTCM and 64KB OCRAM
This commit fixes boot failure with imxrt1050x/nxp_demo configuration.
Boot failure happening due to improper FlexRAM configuration in case of
DTCM usage. It is mandatory to have 64KB FlexRAM allocated for OCRAM at
all times. Due to this max size of DTCM is limited to 256KB. Issue was
happening because we were configuring complete 512KB FlexRAM as DTCM.
Change this configuration to reserve 64KB for OCRAM and 256KB for DTCM.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>