imxrt/semc: make a dependency with C99_BOOL8 and cover it with CONFIG
1.
Configuration of SEMC is not working with source code.
There are some configurations in Kconfig of imxrt but they are not used.
Let's cover semc source files with appropriatary configs.

2.
To use SEMC, CONFIG_C99_BOOL8 should be turned on.
Let's make a dependency with it to enable it automatically.

3.
SDRAM's start address is 0x8000 0000 and size is 33554432.
They are written in imxrt_semc_sdram_init() of imxrt_semc_sdram.c file.
Let's set them as default values of configs.

Signed-off-by: sunghan-chang <sh924.chang@samsung.com>